dt-bindings: pdc: Add SPI config register
In addition to configuring the PDC, additional registers that interface
the GIC have to be configured to match the GPIO type. The registers on
some QCOM SoCs are access restricted, while on other SoCs are not. They
SoCs with access restriction to these SPI registers need to be written
from the firmware using the SCM interface. Add a flag to indicate if the
register is to be written using SCM interface.

Change-Id: I89ba52ded2451071973b81c83a53964d7ec2e486
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>