ARM: dts: msm: Make the CDSP secure carveout mappable on Lahaina
The ION carveout heap manages the carveouts under its domain
by relying on the assumption that the kernel is aware of this
memory. When the memory is marked as "no-map", the kernel
does not maintain any information about the memory, which
prevents the carveout heap from managing the memory. Thus,
allow for the carveout heap memory to be mapped by the kernel.

Change-Id: I6516d2cd8fcacdb975a3190254e79bbce51e1d24
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>