ARM: dts: msm: Fix the gem_noc node address and reg size
The address that is part of the DT node name is incorrect. It has no
functional impact, but this name is used in several places it might
be confusing for someone inspecting the DT.

While looking at this, the register range seems to be different from
what's in the docs, so fix this too.

Change-Id: I50710ce26addc4b5ba1f4bb679ab91f21daf6bd1
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>