ARM: dts: msm: Update disp_bcm_voter to only wait for AMC
Normally, the TCS needs to wait for completion in both AMC and WAKE
sets, since we need to be sure all necessary system resources are
available before they are used. However, the display RSC has strict
wakeup timelines to meet and thus cannot take the latency hit caused by
waiting for completion in the WAKE set. The display RSC will perform
additional processing in between when the WAKE request is sent and
completed. The apps RSC will hold minimum votes on each required
resource to ensure they are enabled, which prevents any unclocked
access. The display RSC is just responsible for scaling to the required
performance levels. The worst that can happen if the WAKE request isn't
satisfied in time is a dropped frame.

Change-Id: Id3757b2ffc73ee93ed8f97d7d591e135150c9cdc
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>