ARM: dts: msm: Set no-status-check-on-disable for Lahaina PCIe GDSCs
The Lahaina PCIe GDSCs are votable between HLOS and ADSP, so the GDSC
may not turn off when HLOS removes its vote since ADSP may be voting as
well. Thus, don't poll for the GDSC to turn off. Otherwise,
gdsc_disable() will return an error when ADSP is voting them on.

Change-Id: I1472843e8bdfa0344abf13e76156f9c0d402885c
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>