ARM: dts: qcom: reduce PM8008J LDO min voltages for Lahaina
Reduce the minimum allowed voltage for PM8008J LDO 3, 4, and 7 so
that the new minimum voltage + the dropout voltage is less than
or equal do the minimum allowed voltage of each LDO's supply
regulator.  This ensures that the supply regulators can be scaled
down to their own minimum voltage when the PM8008J LDOs are not
in use.

Change-Id: I3cb125ff9335d7f9842981351677bc42235b2627
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>