ARM: dts: msm: add cmd to vid switch commands for qsync panel
Add DCS commands that would be sent during a panel operating
mode switch from command mode to video mode.
The switch command sets the full frame width and height as the
horizontal address and vertical address for the DDIC on switching to
video mode. If this is not done and the switch occurs during a partial
frame update corruption can be observed post switch.

Change-Id: I0b12eb24e4fdac5b0e4bbf42493d8bc0081cc361
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>