ARM: dts: qcom: Remove msm watchdog node properties
Due to the creation of a watchdog framework the bark time,
pet time, wakeup-enable, ipi-ping and userspace pet enable
properties have moved to configs in the kernel instead of
as properties of the device. This enables the ability to
set these without the need to change them in the device
tree.

Change-Id: I738fae367aeded48a10aa56e08846ffc8c352bea
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>