ARM: dts: msm: Add the mem-buf device for Shima
Add the mem-buf device for shima to enable memory transfers
between VMs.
The role of the mem-buf driver on the primary VM is to satisfy
allocation requests. There is no realistic scenario in which it
would need to allocate memory from other VMs, so mark the mem-buf
device as a supplier for the primary VM.

Change-Id: Id675fbe6d065063227a518ba66a275910fd12e3e
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>