ARM: dts: msm: Remove msm watchdog node properties for Shima
Due to the creation of a watchdog framework the bark time,
pet time, wakeup-enable, ipi-ping and userspace pet enable
properties have moved to configs in the kernel instead of
as properties of the device. This enables the ability to
to set these without the need to change them in the device
tree.

Change-Id: I65b24de8003bca72113ce5abdd18ebbaff9c1152
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>