ARM: dts: msm: Modify the "init-voltage" of PON rails for shima
Setting the "init-voltage" of the PON rails less than the default
voltage leads to momentary voltage dip on the regulator output.
To avoid this behaviour, set the "init-voltage" to a default
recommended value. Also add ACTIVE_ONLY and SLEEP_ONLY nodes for
LDO1C and LDO10C as per the clients requirement to vote for
ACTIVE and SLEEP votes.

Change-Id: I1e83e47a97cc013bbc1d5bc8c48cbe253b5932eb
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>