ARM: dts: msm: update vig/dma block offset for regdma
When programming vig/dma sub-blocks using REGDMA, the addresses
should be offset with respect to sub-block top. Currently these
addresses are specified with respect to the sspp top instead.
This change adds a new node to capture the relative offset of the
sub-block top for Waipio target.

This change also updates the dma sub-block offsets for Waipio target.

Change-Id: If3775b4ca8cff4ed14d4df3255e256410cab2724
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>