ARM: dts: msm: Add UCSI connector to USB port on Waipio
Add graph nodes representing the USB connector and its remote
endpoints to the UCSI device and the USB controller respectively.
Add 'usb-role-switch' properties to the DWC3 devices and update
its dual-role mode to 'otg'.

This is in order to describe a connection between the pmic_glink
and the DWC3 device to allow the former to notify the USB controller
when to start up in USB host or peripheral mode.

Change-Id: I7e2a9388785bacf17b3240aa939a02c1ec6be0d4
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>