ARM: dts: qcom: Use SPSS_SHARED VMID instead of SPSS
VMID_TZ cannot be hyp-assigned to directly, so it must be removed from
the assign mask for the SP-TZ heap.  Moreover, VMID_CP_SPSS_SP_SHARED
should be used in place of VMID_CP_SPSS_SP when hyp-assigning memory
that is to be shared with TZ (memory will become automatically
available to TZ when hyp-assigned to SPSS_SHARED). Make the needed
updates in the sp-tz carveout heap VMID mask correspondingly.

Change-Id: If539caeae9502a7718e225caa80c04ca02181b88
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>