ARM: dts: msm: Add qcom,max-align for display heap for waipio
add_memory_subsection()/remove_memory_subsection require the base and
size of the memory region to be aligned to 2Mb. These APIs are invoked
when transferring memory to SVM, and currently such allocations come
solely from the display heap.

This change does not attempt to handle the case where a memory region
smaller than 2MB is transferred to SVM.

Change-Id: I7e1fd28a649795c72ea6d8b34886018daa7d72d7
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>