ARM: dts: qcom: Disable VM low power modes for waipio
There are changes related GIC 1:N changes that went recently which is
causing a race condition in hypervisor that lead to VM is entering
into PSCI suspend even there are pending interrupts at GIC.Pending
interrupts cause vCPUs to wake up immediately to serve the interrupts.
This scenario is happening in a tight loop and causing vCPUs to enter
and wakeup uncessarily and lead to RT throttle. This makes VM not boot.
And hence this change to disable VM low power modes for waipio.

Change-Id: I852a57525e10c06f28de17f21775434f5441c834
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>