ARM: dts: msm: use TLMM to reference trusted touch GPIOs on Waipio
Use TLMM gpio pin numbers to lookup the base address offsets of
the GPIO pins used for touch input. This involves adding to the
list of valid GPIOs for TLMM, adding these pins to the list of
IO memory ranges shared between VMs, and converting the hard-
coded offsets to TLMM GPIO references for touch. This change
also adds the missing max-touch-number for MTP and CDP
platforms.

Change-Id: I8d107343acd2908be9d847c72480abd19fda04bd
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>