ARM: dts: msm: waipio HDK HDMI bring up
Add support of HDMI display for waipio HDK, there is
a 'DSI0_SW' switch on HDK main board, for which setting
to 'ON' means show image through HDMI to external display monitor
while setting to 'OFF' means show image to DSI panel.
Users need set switch to 'ON' to turn on HDMI display.

Change-Id: Iaa2e8ae6a0ccd5284b81c1ba64ab69e9e84ce87c
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>