ARM: dts: msm: Update actlr mask setting for diwali
Previously, the actlr settings was derived from a lookup table, where
the entry is found if its sid/mask set overlaps with the set described by
and input sid/mask. Due to these overlaps, several matches may be found,
which is detected as an error condition.
To avoid overlaps smmu driver is updated, to be inligned with it,
updating the actrl mask settings.

Change-Id: Ic7445732b6658c9c353cdfc7303854fc0d155dee
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>