Revert "ARM: dts: msm: add shutdown voltage for Waipio platforms"
This reverts commit 78130a44eb29c1ef853ac8d62a1a4d8751fcb008.

With the commit being reverted, kernel_power_off() would be invoked
whenever the battery SOC reaches 0, battery is discharging and
the battery voltage goes below the specified shutdown voltage. This
was made to avoid an UVLO reset if the shutdown initiated by the
userspace takes longer than expected time. However, in certain
conditions, when filesystem sync didn't go as expected and with UFS
or storage drivers removing the devices in their shutdown callbacks,
init process is getting a SIGBUS exit code ending up in a kernel
panic. With recent testing, it is also found out that an UVLO reset
is not happening as before even though the shutdown still takes
longer than expected. Taking this into consideration, reverting
the change seems fine.

Change-Id: I3c6e22d154583e80570ad8e4c8e24c5bad9060d4
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>