ARM: dts: msm: use TLMM to reference trusted touch GPIOs
Use TLMM gpio pin numbers to lookup the base address offsets of the
GPIO pins used for touch input. This involves adding to the list of
valid GPIOs for TLMM, adding these pins to the list of IO memory
ranges shared between VMs, and converting the hard-coded offsets to
TLMM GPIO references for touch.

Change-Id: Ia6fa949ecde8897e27915f43be6616d2d4b502f6
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>