ARM: dts: msm: Add UCSI connector to USB port on Cape
Add graph nodes representing the USB connector and its remote
endpoints to the UCSI device and the USB controller respectively.
Add 'usb-role-switch' properties to the DWC3 devices.

This is in order to describe a connection between the pmic_glink
and the DWC3 device to allow the former to notify the USB controller
when to start up in USB host or peripheral mode.

Change-Id: I4336f8d6c5c2b3e8ce5efa6fc204fd8ec058c1b9
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>