ARM: dts: msm: Update probe sequence for diwali regulators
VDD_MMCX_LEVEL is not able to register for a proxy vote
if its parent VDD_CX_MMCX_SUPPLY_LEVEL is not probed causing
no proxy vote to be there on it. Fix this by moving its parent
regulator VDD_CX_MMCX_SUPPLY_LEVEL above it in DT sequence, so
that the parent regulator would be probed first.

Change-Id: I557e3418d03996c4882eab9344dca9dd74b87a38
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>