ARM: dts: msm: add stub regulator for amoled ibb supply of Diwali
In the absence of ibb regulator supply node, the display is getting
a dummy regulator for ibb. The dummy regulator driver does not define
the regulator_ops. Because of this regulator_set_mode() call for ibb
regulator is failing (non-fatal).
Use stub regulator for ibb supply to fix this error.

Change-Id: I07b70cd0fdd49f1ee16285d2a65fbad7ca69bcc4
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>