ARM: dts: msm: Make the SCM pseudo-device use CMA for allocations
The allocations using the SCM pseudo-device sometimes need to be kept
in the lower 4 GB of memory in order to be usuable by certain devices.
Accordingly, make the SCM pseudo-device use a CMA area for its
allocations through the dma_alloc_* APIs to ensure its allocations are
kept in the lower 4 GB of memory.

Change-Id: Ia58ba0d97a9ddd3d680e26939837b866fa9b10f3
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>