ARM: dts: msm: Remove min/max voltages for LDO24B for parrot
LDO24B has its voltage set to 2.4V (min voltage) by regulator framework
as part of regulator registration if the initial voltage is not set.
This regulator will only be voted for enabling/disabling
conditions by UFS, so fix this by removing the min/max voltage values
for this regulator.

Change-Id: I13f82e919e3e46f2e1a6cc0905f69f11c6fb2b2d
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>