bindings: leds: Add use-qti-battery-interface DT property
Add new DT property use-qti-battery-interface for ADSP charger
designs, to allow reading the battery properties from
qti-battery-charger driver.This is needed for max current
calculations in the flash driver. If this property is not added,
flash driver will try to read these properties over the IIO interface,
which is used in non-ADSP charger designs.

Change-Id: I16ed4ca49808aa2e1e7fce03152db76f5e7ba562
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>