ARM: dts: msm: Device tree support for muticard tray removal FR
Add L24B_PBS and L28B_PBS regulators to support multicard
tray removal FR84471 for ravelin which may be needed for all the
chipsets where PMIC doesn't have PBS ram sequence to turn off the
regulators asap the tray is removed.

Need to vote for enable/disable on the above new handles provided
and rest for voltage and mode need to vote on old handles.

Change-Id: I862e792eeb89fb13ad6c4181f6761229f5b2fe39
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>