bindings: Update DT-binding property for core hang detect
Binding properties "qcom,threshold-arr" and "qcom,config-arr"
assumes that logical CPU number & physical CPU has 1:1 mapping
which is not true when any CPU is not available.

To resolve this we are replacing old properties with a new property
"qcom,chd-percpu-info" and map core-hang registers wrt to CPU phandles.
The format is <&CPUx_Phandle CPUx_Threshold CPUx_Config>.

Change-Id: Ib5ce47b5969e1203e6508e6c77dfdc8b6d7de65d
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>