Revert "ARM: dts: msm: Enable spm level 5 for ufs"
This reverts commit 376dcc687779a63285a87570985adb7d6c130343.

This change leading to stability issue where rpmb reads are failing.
To mitigate, huge design change is required which is not feasible at
this stage. Hence reverting this change.

Moreover going forward, we need to follow spm-lvl = 3 only, as per
upstream recommendation, also this is inline with PT targets as it
helps reduce resume latency.

Change-Id: I18fc5323032e586957ffbf97e5f736c7ce54eb72
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>